Fixed compilation when targeting x64
- The monkey in the Sound sample now faces the camera

+++ Text/Game.h
@@ -42,7 +42,7 @@ public:
 	ConfigTable& config;
 
 	std::vector<Font::Ptr> fonts;
-	unsigned currentFont = 0;
+	size_t currentFont = 0;
 	bool direction = false;
 	float angle = 0.0f;
 };
